remove another property; refactor
