+++ airbyte-cdk/bulk/core/load/src/testFixtures/kotlin/io/airbyte/cdk/load/test/util/destination_process/DestinationProcess.kt
@@ -6,6 +6,7 @@ package io.airbyte.cdk.load.test.util.destination_process
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings
 import io.airbyte.cdk.command.FeatureFlag
+import io.airbyte.cdk.load.command.Property
 import io.airbyte.cdk.load.test.util.IntegrationTest
 import io.airbyte.protocol.models.v0.AirbyteMessage
 import io.airbyte.protocol.models.v0.ConfiguredAirbyteCatalog
@@ -99,19 +100,3 @@ abstract class DestinationProcessFactory {
             }
     }
 }
-
-/**
- * Represents a micronaut property, which has a corresponding entry in micronaut's `application.yml`
- * file, which is populated by an environment variable. Just a pair of the micronaut property name,
- * and that corresponding env var name.
- *
- * For example, this application.yaml:
- * ```yaml
- * airbyte:
- *   destination:
- *     foo-bar: ${FOO_BAR}
- * ```
- *
- * Would be represented as `Property("airbyte.destination.foo-bar", "FOO_BAR")`.
- */
-data class Property(val micronautProperty: String, val environmentVariable: String)
