+++ airbyte-cdk/bulk/toolkits/load-aws/build.gradle
@@ -1,6 +1,4 @@
 dependencies {
     implementation project(':airbyte-cdk:bulk:core:bulk-cdk-core-base')
     implementation project(':airbyte-cdk:bulk:core:bulk-cdk-core-load')
-
-    testFixturesImplementation testFixtures(project(':airbyte-cdk:bulk:core:bulk-cdk-core-load'))
 }
